+++ packages/entities/entities-shared/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.57.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-shared@3.57.1...@kong-ui-public/entities-shared@3.57.2) (2026-07-22)
+
+**Note:** Version bump only for package @kong-ui-public/entities-shared
+
+
+
+
+
 ## [3.57.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-shared@3.57.0...@kong-ui-public/entities-shared@3.57.1) (2026-07-21)